pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-quant"
version = "0.1.0"
description = "LLM-powered paper trading system where Claude acts as portfolio manager"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "45ck" }]
keywords = ["llm", "quant", "paper-trading", "claude", "finance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "anthropic>=0.39.0",
    "duckdb>=1.1.0",
    "polars>=1.0.0",
    "yfinance>=0.2.40",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.9.0",
    "tomli>=2.0.0;python_version<'3.11'",
    "jinja2>=3.1.0",
    "scipy>=1.12.0",
    "pyyaml>=6.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]
dashboard = [
    "streamlit>=1.38.0",
    "plotly>=5.24.0",
]

[project.scripts]
pq = "llm_quant.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_quant"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

# ── noslop quality gates ──────────────────────────────────────────────

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "D",      # docstring rules — enable when the project matures
  "ANN",    # type annotation rules — use mypy for this
  "COM812", # trailing comma (conflicts with formatter)
  "ISC001", # implicit string concat (conflicts with formatter)
  "S101",   # assert — used intentionally in production for invariants
  "PLC0415",# deferred imports — CLI pattern to speed up startup
  "PLR2004",# magic values — acceptable in financial domain (percentages, thresholds)
  "RUF001", # ambiguous unicode — false positives on currency symbols
  "RUF002", # ambiguous unicode in docstrings
  "RUF003", # ambiguous unicode in comments
  "EM102",  # f-string in exception — readability over stack traces
  "TRY003", # raise vanilla args — over-abstraction for a small codebase
  "S608",   # SQL injection — false positive on parameterized queries with placeholders
  "FBT001", # boolean positional args — acceptable for simple flags
  "FBT003", # boolean positional value in call
  "S311",   # non-crypto random — used for non-security purposes
  "TC001",  # typing-only imports — adds complexity for little benefit here
  "TC002",  # typing-only third-party imports
  "TC003",  # typing-only stdlib imports
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
  "S101",    # asserts are the whole point of tests
  "PLR2004", # magic values in test assertions are fine
  "PT018",   # composite assertions — acceptable in tests
  "PLR0913", # test helpers may have many args for flexibility
  "ARG002",  # test stubs may not use all ABC args
  "SLF001",  # tests need to access private members
  "ERA001",  # section dividers in test files are useful
  "PLR0133", # constant comparisons in tests are intentional assertions
  "PLC0206", # dict iteration patterns fine in tests
  "PERF401", # performance not critical in tests
]
"scripts/**/*.py" = [
  "INP001",  # scripts are standalone, not packages
  "T201",    # scripts use print() for user-facing output
  "E501",    # analysis scripts allow long lines for readability
  "BLE001",  # broad exception catching fine in analysis scripts
  "PTH123",  # open() fine in scripts
  "PLR0913", # analysis scripts may have many CLI args
  "C901",    # walk-forward loops are inherently complex
  "ARG001",  # scripts may have unused args for interface consistency
  "ARG002",  # inline classes may not use all ABC args
  "RUF059",  # unused unpacked vars fine in scripts
  "FBT002",  # boolean defaults fine in script helpers
]
"src/llm_quant/arb/**/*.py" = [
  "BLE001",  # broad exception catching is intentional in API parsing code
  "FBT002",  # boolean defaults fine in API client config methods
  "PLR0913", # data ingestion functions may have many parameters
  "TRY401",  # logger.exception with exc arg fine in error handlers
  "T201",    # print_report() is an intentional display utility — print() is correct
]
"src/llm_quant/backtest/engine.py" = [
  "PLR0913", # engine.run() needs many params by design
  "PLR0912", # backtest loop has many branches by nature
  "C901",    # backtest loop is inherently complex
  "PLR0915", # backtest loop has many statements
]
"src/llm_quant/backtest/strategies.py" = [
  "PLR0912", # regime strategy has many conditional branches
  "C901",    # regime strategy is inherently complex
  "ARG002",  # ABC contract requires all params even if unused
]
"src/llm_quant/backtest/strategy.py" = [
  "ARG002",  # reference impl may not use all ABC params
]
"src/llm_quant/backtest/robustness.py" = [
  "N806",    # T, N, S are standard mathematical notation (PBO/CPCV)
  "PLR0913", # run_robustness_gate needs all gate inputs
]
"src/llm_quant/backtest/metrics.py" = [
  "FBT002",  # annualize=True is a clear named boolean
]
"src/llm_quant/db/integrity.py" = [
  "PLR0913", # compute_trade_hash needs all trade fields for hash input
]

[tool.ruff.lint.mccabe]
max-complexity = 13

[tool.ruff.lint.pylint]
max-args = 6
max-branches = 15
max-statements = 65
max-returns = 6
max-bool-expr = 4

[tool.mypy]
strict = true
ignore_missing_imports = true
